requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.1
django-cache-url==3.4.5
django-gravatar2==1.4.4
django-simple-history==3.7.0
environs==9.5.0
idna==3.7
marshmallow==3.20.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
xmltodict==0.13.0
